+++ tests/conftest.py
@@ -4,6 +4,8 @@ import pytest
 
 from ts_backend_check.cli.main import config
 
+# MARK: Invalid
+
 invalid_django_models = config["invalid_model"]["backend_model_path"]
 invalid_ts_interfaces = config["invalid_model"]["ts_interface_path"]
 
@@ -18,8 +20,14 @@ def return_invalid_ts_interfaces():
     return str(invalid_ts_interfaces)
 
 
+# MARK: Valid
+
 valid_django_models = config["valid_model"]["backend_model_path"]
 valid_ts_interfaces = config["valid_model"]["ts_interface_path"]
+valid_check_blank_models = config["valid_model"]["check_blank_model_fields"]
+valid_backend_to_ts_conversions = config["valid_model"][
+    "backend_to_ts_model_name_conversions"
+]
 
 
 @pytest.fixture
@@ -30,3 +38,13 @@ def return_valid_django_models():
 @pytest.fixture
 def return_valid_ts_interfaces():
     return str(valid_ts_interfaces)
+
+
+@pytest.fixture
+def return_valid_check_blank_models():
+    return valid_check_blank_models
+
+
+@pytest.fixture
+def return_valid_backend_to_ts_conversions():
+    return valid_backend_to_ts_conversions